Update Readme with build instructions and credits
Lets make it clear how to build this and add credits for everyone that has helped make this a reality. 

If I forgot you or lost authorship on work you did, I apologize and would welcome being pointed to said commit you did that I can pull authorship from, or if you just want credited here in the readme.